% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qleTest.R
\name{checkMultRoot}
\alias{checkMultRoot}
\title{Assess plausibility of parameter estimates}
\usage{
checkMultRoot(est, par = NULL, opts = NULL, verbose = FALSE)
}
\arguments{
\item{est}{object of class \code{qle}, the estimation results from function \code{\link{qle}}}

\item{par}{list or matrix of estimated parameters as roots of the quasi-score vector}

\item{opts}{list of upper bounds for a root of the quasi-score vector, see details}

\item{verbose}{logical, \code{TRUE} for intermediate output}
}
\value{
A data frame with columns named corresponding to each component of the investigated parameter,
 `\code{minor}`, `\code{det}`, `\code{value}`, `\code{score_max}`, `\code{logdetI}`, `\code{lamI_min}`,
 `\code{lamIm_max}` and `\code{varS_trace}` (see vignette for details). The first column shows the leading minor of
  the observed QI matrix which is not positive definite. Estimated model parameters for which the observed QI matrix
  is not positive definite are excluded from further root selection.
}
\description{
Check out and compare estimated roots of the quasi-score vector
}
\details{
Only for the quasi-likelihood approach the function inspects the (numerical) consistency of the found
 parameters in `\code{par}` by comparing each observed quasi-information matrix with the expected one.
 The degree of dissimilarity of both matrices is measured by certain scalar equivalent criteria (see vignette)
 and the parameter for which these are smallest is chosen.
}
\examples{
 data(qleresult)

 # and just a single root 
 checkMultRoot(OPT,verbose=TRUE)

}
\author{
M. Baaske
}
