% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsOpt.R
\name{getDefaultOptions}
\alias{getDefaultOptions}
\title{Show default options}
\usage{
getDefaultOptions(xdim)
}
\arguments{
\item{xdim}{size of problem dimension (number of model parameters to estimate)}
}
\value{
List of options.
}
\description{
Print all default options for optimization
}
\details{
The function returns a lists of all available options
 for functions \code{\link{qscoring}} and \code{\link{qle}}.
}
\examples{
getDefaultOptions(xdim=2)
 
}
\author{
M. Baaske
}
