% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsOpt.R
\name{globalSearch}
\alias{globalSearch}
\title{Find new global evaluatino point}
\usage{
globalSearch(x0, qsd, w, method = c("nloptr", "direct"), opts = list(),
  control = list(), ..., obs = NULL, optInfo = FALSE, check = TRUE,
  pl = 0L, verbose = FALSE)
}
\arguments{
\item{x0}{(named) numeric vector, the starting point}

\item{qsd}{object of class \code{\link{QLmodel}}}

\item{w}{scalar value, cycling weight}

\item{method}{names of possible minimization routines (see details)}

\item{opts}{list of control arguments passed to \code{\link[nloptr]{nloptr}} routines defined in \code{method} for global search of maximum and minimum value of quasi-deviance}

\item{control}{list of control arguments passed to \code{\link[nloptr]{nloptr}} for maximization of selection criterion}

\item{...}{further arguments passed to \code{\link{covarTx}}}

\item{obs}{numeric vector of observed statistics, overwrites `\code{qsd$obs}`}

\item{optInfo}{logical, \code{FALSE} (default), not yet used argument (ignored)}

\item{check}{logical, \code{TRUE} (default), whether to check input arguments}

\item{pl}{numeric value (>=0), the print level}

\item{verbose}{if \code{TRUE} (default), print intermediate output}
}
\value{
list which contains the new evaluation point and the value of the criterion function
}
\description{
The function finds the maximum value of the global selection criterion, which basically
 is a distance weighted version of the quasi-deviance over the whole parameter domain.
}
\seealso{
\code{\link{searchMinimizer}}
}
