% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige.R
\name{Select subset of statistics}
\alias{Select subset of statistics}
\alias{optStat}
\title{Optimal selection of a number of statistics}
\usage{
optStat(theta, qsd, kmax, ..., cl = NULL, verbose = FALSE)
}
\arguments{
\item{theta}{list or matrix of points where to compute the criterion function
and to choose the `\code{kmax}` best statistics given the QL model `\code{qsd}`}

\item{qsd}{object of class \code{\link{QLmodel}}}

\item{kmax}{number of (optimal) statistics to be selected}

\item{...}{further arguments passed to \code{\link{quasiDeviance}} or \code{\link{mahalDist}}}

\item{cl}{cluster object, \code{NULL} (default), of class \code{MPIcluster}, \code{SOCKcluster}, \code{cluster}}

\item{verbose}{logical, \code{TRUE} for intermediate output}
}
\value{
A list which consists of 
	\item{id}{ indices of corresponding statistics}
	\item{names}{ names of statistics (if provided)}
 \item{prop}{ proportion of contribution of each selected statistic to each parameter} 
 \item{sorted}{ list of statistics (for each parameter) sorted in decreasing order of contribution to the quasi-information}
}
\description{
The function finds a subset of statistics of size at most equal to `\code{kmax}`
which are optimal in the sense of highest contributions to minimize the expected estimation error
of the model parameter using the quasi-information.
}
\examples{
 data(normal)
 optStat(c("mu"=2,"sigma"=1),qsd,kmax=2)[[1]]

}
\author{
M. Baaske
}
