% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qleTest.R
\name{print.cvTest}
\alias{print.cvTest}
\title{print \code{cvTest} results}
\usage{
\method{print}{cvTest}(res, pl = 1, digits = 5, format = "e", ...)
}
\arguments{
\item{pl}{ignored}

\item{digits}{number of (significant) digits}

\item{format}{format character(s), see \code{\link{formatC}}}

\item{...}{ignored}

\item{x}{object of class \code{cvTest} from \code{\link{rootCVtest}}}
}
\description{
print the results from \code{\link{cvTest}}
}
