% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsOpt.R
\name{print.QSResult}
\alias{print.QSResult}
\title{print results of class \code{QSResult}}
\usage{
\method{print}{QSResult}(x, pl = 1L, digits = 4, format = "e", ...)
}
\arguments{
\item{x}{object of type \code{QSResult} from \code{\link{qscoring}}}

\item{pl}{numeric positive value, the print level (higher values give more information)}

\item{digits}{number of digits to display}

\item{format}{format character(s), see \code{\link{formatC}}}

\item{...}{ignored, additional arguments}
}
\description{
S3 method to print the results from \code{\link{qscoring}}.
}
