% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qleTest.R
\name{qDPress}
\alias{qDPress}
\title{Cross-valiation based variance and CI}
\usage{
qDPress(points, values, cvm, ..., jack = FALSE,
  fun = getOption("qle.multicore", "mclapply"), cl = NULL, pl = 0L,
  verbose = FALSE)
}
\arguments{
\item{points}{object of class \code{QSResult}}

\item{values}{object of class \code{\link{QLmodel}}}

\item{cvm}{object}

\item{...}{arguments passed to the simulation function `\code{sim}`, \code{\link{searchMinimizer}} and \code{\link{multiSearch}}}

\item{jack}{logical, \code{FALSE} (default)}

\item{cl}{cluster object}

\item{pl}{numeric value (>=0), the print level}

\item{verbose}{if \code{FALSE} (default), print intermediate output}
}
\value{
list of
}
\description{
Get CV based
}
\details{
The function
}
\author{
M. Baaske
}
