% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{QLmodel}
\alias{QLmodel}
\title{Construct the quasi-likelihood approximation model}
\usage{
QLmodel(qldata, lb, ub, obs, mods, nfit = 1, cv.fit = FALSE,
  var.type = c("cholMean", "wcholMean", "logMean", "wlogMean", "kriging",
  "logKrig", "const"), krig.type = "var", criterion = c("qle",
  "mahal"), verbose = FALSE)
}
\arguments{
\item{qldata}{data frame of (initial) simulation results (see \code{\link{setQLdata}})}

\item{lb}{numeric vector of lower bounds defining the (hyper)box}

\item{ub}{numeric vector of upper bounds defining the (hyper)box}

\item{obs}{numeric vector of observed statistics}

\item{mods}{list of (fitted) covariance models (see \code{\link{fitSIRFk}})}

\item{nfit}{number of cycles, \code{nfit=1} (default), after which covariance
parameters are re-estimated and otherwise only re-used}

\item{cv.fit}{logical, \code{FALSE} (default), whether to re-fit CV models (re-estimate covariance parameters)}

\item{var.type}{name of the variance approximation method (see \code{\link{covarTx}})}

\item{krig.type}{name of kriging variant used, either "\code{var}" for kriging with computation of krigin prediction variance or "\code{dual}"}

\item{criterion}{global criterion function for sampling and minimization, either "\code{qle}" or "\code{mahal}"}

\item{verbose}{logical, \code{FALSE} (default), whether to give intermediate output}
}
\value{
An object of class \code{\link{QLmodel}} which stores the data frame of simulation results, bounds on
 the parameter space, covariance models for kriging, vector of observed statistics as well as options for
 kriging and covariance parameter estimation.
}
\description{
Aggregate and construct the data for quasi-likelihood estimation
}
\details{
The function aggregates all required information for quasi-likelihood estimation and defines the input object to the 
  function \code{\link{qle}}, stores the fitted covariance models of the sample means of the statistics and the type of variance
  matrix approximation. For an advanced setup of the estimation procedure and more involved statistical models this function
  explicitly offers the data structure to construct individual covariance models for each statistic, see \code{\link{setCovModel}}.
	 The kriging procedure includes the computation of prediction variances based on kriging.
}
\examples{

data(normal)

# As an example we re-use the stored normal data and fit 
# a generalized covariance model to the data using simulation
# variances as local variances for REML estimation.
mods <- fitSIRFk(qsd$qldata, verbose=TRUE)

# construct QL approximation model
qsd <- QLmodel(qsd$qldata,qsd$lower,qsd$upper,
			    c("T1"=2,"T2"=1),mods)

}
\seealso{
\code{\link{getQLmodel}}, \code{\link{updateCovModels}}
}
\author{
M. Baaske
}
