% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qleTest.R
\name{rootCVtest}
\alias{rootCVtest}
\title{Cross-valiation based testing}
\usage{
rootCVtest(est, par0, qsd, cvm, ..., alpha = 0.05, type = c("qle",
  "mahal", "quasiWald"), na.rm = TRUE, check = TRUE, info = TRUE,
  cl = NULL, pl = 0L, verbose = FALSE)
}
\arguments{
\item{est}{object of class \code{QSResult}}

\item{par0}{optional, vector of parameters or list of parameters for testing the null hypothesis}

\item{qsd}{object of class \code{\link{QLmodel}}}

\item{cvm}{list of prefitted covariance models obtained from function \code{\link{prefitCV}}}

\item{...}{arguments passed to the simulation function `\code{sim}`, \code{\link{searchMinimizer}} and \code{\link{multiSearch}}}

\item{alpha}{significance level for testing the hypothesis}

\item{type}{name of test statistic, \code{"qle"} (default), either "\code{qle}" or "\code{mahal}" or "\code{quasiWald}" which overwrites the function criterion}

\item{na.rm}{logical, \code{TRUE} (default), whether to remove `NA` values from the matrix of re-estimated parameters}

\item{check}{logical, \code{TRUE} (default), whether to check input arguments}

\item{info}{logical, \code{TRUE} (default), compute additional error matrices}

\item{cl}{cluster object, \code{NULL} (default), of class \code{MPIcluster}, \code{SOCKcluster}, \code{cluster}}

\item{pl}{numeric value (>=0), the print level}

\item{verbose}{if \code{FALSE} (default), print intermediate output}
}
\value{
An object of class \code{qleTest} (see \code{\link{qleTest}})
}
\description{
A cross-validation based approach for testing an approximate root
}
\details{
The function uses a cross-validation approach to test an approximate root being a 
 reasonable estimate of the parametric model. Note that this approach only accounts for the error 
 in the approximation model of the true quasi-score function by kriging because model observations
 are fixed.
}
\author{
M. Baaske
}
