% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsOpt.R
\name{summary.dsample}
\alias{summary.dsample}
\title{Generating Basic Summary Statistics of Marginal Distributions}
\usage{
\method{summary}{dsample}(object, n = 5, ...)
}
\arguments{
\item{object}{a \code{data.frame}, contains the sample drawn via either the Fu-Wang algorithm or the Wang-Lee algorithm}

\item{n}{the first n samples}

\item{...}{more arguments}
}
\description{
Producing basic summary statistics (the mean, the standard deviation and the first five modes) from the sample drawn via either the Fu-Wang algorithm or the Wang-Lee algorithm, for all marginal distributions of the target distribution.
}
\author{
Chel Hee Lee \email{chl948@mail.usask.ca}, Liqun Wang \email{liqun.wang@umanitoba.ca}
}
