% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsOpt.R
\name{traceQIconstr}
\alias{traceQIconstr}
\title{Find a new local evaluation point}
\usage{
traceQIconstr(x0, qsd, b, method = c("nloptr"), opts = list(), ...,
  optInfo = FALSE, check = TRUE, pl = 0L, verbose = FALSE)
}
\arguments{
\item{x0}{(named) numeric vector, the starting and center point of the feasible region}

\item{qsd}{object of class \code{\link{QLmodel}}}

\item{b}{upper bound on either quasi-deviance or mahalanobis distance criterion function values as an upper bound constraint (as a level set function) which defines the feasible region}

\item{method}{names of possible minimization routines, currently only "\code{nloptr}" is supported (see details)}

\item{opts}{list of control arguments passed to \code{\link[nloptr]{nloptr}} routines defined in \code{method} for global search of maximum and minimum value of quasi-deviance}

\item{...}{further arguments passed to \code{\link{covarTx}}}

\item{optInfo}{logical, \code{FALSE} (default), not yet used argument (ignored)}

\item{check}{logical, \code{TRUE} (default), whether to check input arguments}

\item{pl}{numeric value (>=0), the print level}

\item{verbose}{if \code{TRUE} (default), print intermediate output}
}
\value{
list which contains the new evaluation point as the maximizer and the corresponding value of the criterion function
}
\description{
The function maximizes the total prediction variance of the quasi-score vector in the vicinity of
 the center point `\code{x0}` (usually the current parameter estimate). More specifically, 
 a distance weighted version of the trace of the quasi-information matrix is maximized over a local region defined 
 by a level set of hight `\code{b}` w.r.t. the criterion function (either quasi-deviance or Mahalanobis distance)
 based on the observed statistics. The function can be used to find a suitable next evaluation point. As one of the
 global selection criteria the function can be called by \code{\link{qle}} to find a new evaluation point if testing
 approximate roots is enabled.
}
\seealso{
\code{\link{searchMinimizer}}
}
\author{
M. Baaske
}
