% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsOpt.R
\name{traceQIconstr}
\alias{traceQIconstr}
\alias{traceQIcontr}
\title{Find new local evaluation point}
\usage{
traceQIcontr(x0, qsd, b, method = c("nloptr"), opts = list(), ...,
  optInfo = FALSE, check = TRUE, pl = 0L, verbose = FALSE)
}
\arguments{
\item{x0}{(named) numeric vector, the starting point}

\item{qsd}{object of class \code{\link{QLmodel}}}

\item{b}{upper bound on either quasi-deviance or mahalanobis distance as a constraint defining the feasible region}

\item{method}{names of possible minimization routines, currently only "\code{nloptr}" is supported (see details)}

\item{opts}{list of control arguments passed to \code{\link[nloptr]{nloptr}} routines defined in \code{method} for global search of maximum and minimum value of quasi-deviance}

\item{...}{further arguments passed to \code{\link{covarTx}}}

\item{optInfo}{logical, \code{FALSE} (default), not yet used argument (ignored)}

\item{check}{logical, \code{TRUE} (default), whether to check input arguments}

\item{pl}{numeric value (>=0), the print level}

\item{verbose}{if \code{TRUE} (default), print intermediate output}
}
\value{
list which contains the new evaluation point and the value of the criterion function
}
\description{
The function finds the maximum value of the local selection criterion, which basically
 is a distance weighted version of the trace of the quasi-information matrixe over a local region defined 
 by the level set of hight \code{b} w.r.t. the quasi-deviance using the actual observed statistics.
}
\seealso{
\code{\link{searchMinimizer}}
}
\author{
M. Baaske
}
