% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsOpt.R
\name{traceQIsamp}
\alias{traceQIsamp}
\title{Finding a new evaluation point}
\usage{
traceQIsamp(x0, qsd, QD, Sig0 = NULL, nsim = 1000, type = 0L,
  fit = FALSE, pmin = 0.25, ..., optInfo = FALSE, cl = NULL,
  pl = 0L, verbose = FALSE)
}
\arguments{
\item{x0}{numeric (named) vector or list as the center point of sampling region,
e.g., an estimate of the model parameter}

\item{qsd}{object of class \code{\link{QLmodel}}}

\item{QD}{list of criterion function results, either from calling \code{\link{quasiDeviance}} or \code{\link{mahalDist}}}

\item{Sig0}{sample covariance matrix of parameters given in \code{QD}, default \code{NULL}, which is then computed}

\item{nsim}{sample size of random points for computing the integral}

\item{type}{integer, \code{type=0} (default), use uniform sampling of points, \code{type=1} multivariate normal random points}

\item{fit}{logical, \code{FALSE} (default), whether to re-fit kriging covariance models of the statistics}

\item{pmin}{minimum required ratio of points falling inside the overall design region (hyperbox of parameters)}

\item{...}{optional, arguments passed to \code{\link{quasiDeviance}}, center point '\code{theta}' and '\code{W}' as weighting matrix for kernel estimate of covariance matrix,
cross-validation models '\code{cvm}', see \code{\link{quasiDeviance}} for details}

\item{optInfo}{logical, \code{FALSE} (default), not yet used argument (ignored)}

\item{cl}{cluster object, \code{NULL} (default), of class \code{MPIcluster}, \code{SOCKcluster}, \code{cluster}}

\item{pl}{print level, use \code{pl}>0 to print intermediate results}

\item{verbose}{if \code{TRUE} (default), print intermediate output}
}
\value{
list of the criterion function value corresponding to the model parameter which minimizes
 the integrated total prediction variance of the quasi-score vector
}
\description{
Minimizing the average total prediction variance of the quasi-score vector
}
\details{
Based on the list of sampling candidates \code{QD} one point at a time is (sequentially) added to
the current sampling design. The integrated (averaged) total prediction variance of the quasi-score vector is then
computed based on this new design over a randomly generated set of candidate evaluation points with a sample covariance
derived from all sampled candidates. The integral is taken over a discrete set of points either generated uniformly within
the design region defined by simple bound constraints if '\code{Sig0}' is \code{NULL} or multivariate normally distributed points
with covariance matrix '\code{Sig0}' and mean parameter vector '\code{theta0}'.
}
\seealso{
\code{\link{quasiDeviance}} 
#
}
\author{
M. Baaske
}
