% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qleTest.R
\name{checkMultRoot}
\alias{checkMultRoot}
\title{Inspect estimated parameters}
\usage{
checkMultRoot(est, par = NULL, verbose = FALSE)
}
\arguments{
\item{est}{object of class \code{qle}, the estimation results from function \code{\link{qle}}}

\item{par}{list or matrix of estimated parameters as roots of the quasi-score vector}

\item{verbose}{logical, \code{TRUE} for intermediate output}
}
\value{
A data frame with columns named corresponding to each component of the investigated parameter,
 `\code{quasi-deviance}`, `\code{Minor}`, `\code{det}`, `\code{max}`,
 `\code{trace}` and `\code{score_max}` (see vignette). The second column shows the leading minor of
  the observed QI matrix which is not positive definite. If so, then the corresponding parameter estimate
  cannot be consistent at least in theory. The rows show the corresponding values for each parameter passed by
  `\code{par}`.
}
\description{
Check out and compare estimated roots of the quasi-score vector
}
\details{
Ony in case of the quasi-likelihood approach using the criterion function "\code{qle}" we can compare the
 observed quasi-information matrix with the expected one by some (scalar equivalent) criteria (see vignette).
 We measure this dissimilarity of both matrices evaluated at some given approximate roots `\code{par}` for which these
 criteria should be neglectably small to be a plausible root. By the smallest value we extract the best root.
}
\examples{
 data(normal)
 # a dummy estimation result: no new evaluations
 OPT <- qle(qsd,qsd$sim,global.opts=list("maxeval"=0))
 # and just a single root 
 checkMultRoot(OPT,verbose=TRUE)

}
\author{
M. Baaske
}
